[REFACTOR] #35 - move notificationViewModel into fragment package
